Current version of pisabell2;  Changes: 1) in Devialet command and update packets; and 2) added mqtt interface
Changes in Devialet update packets: critical status bytes appear further down the packet. Is this due to the fact that I am using a D220 PRO rather than a D200? Or the fact that I am chaining several Devialet amps?
Changes in Devialet command packets: the code for determing the two bytes associated with each source was not working. Has been replaced with hard coded bytes spelled out in dictionaries.
The mqtt code is working but probably provisional because of my substandard Python writing abilities. I am using the Mosquitto broker running on the same computer as Devimote; I have not tested the code with a different broker.
Ideally, the user might be given the choice to use only a gui, only mqtt, or both.